+++ packages/react-components/src/components/SummaryList/index.ts
@@ -1,0 +1,7 @@
+export { SummaryList } from './SummaryList';
+export type {
+  SummaryListActionItem,
+  SummaryListContent,
+  SummaryListProps,
+  SummaryListRow,
+} from './SummaryList';